requirements.txt
astroid==2.4.1
CoAPthon3==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
protobuf==3.12.2
pylint==2.5.2
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
